create vikings.ts and vikings.test.ts file
